added view and refactored
