+++ templates/all_items.html
@@ -10,11 +10,11 @@
     <div class="col-md-12">
       <a href="{{url_for('create_item_page')}}"> Click here to add new card to ToDo </a>
       <h2>Current To Do Items from Trello:</h2>
-        {% for todo in cards_from_todo %}
+        {% for todo in todos.to_do_items %}
           <form action="{{url_for('Items_Done')}}" method="POST">
             <input type ='text' name='card_name' value="{{todo.title}}" readonly="readonly">
-            <input type = 'text' name="list_id" value="{{todo.list_id}}" readonly="readonly">
-            <input type = 'text' name ="card_id" value="{{todo.id}}" readonly="readonly">
+            <input type = 'text' name="status" value="{{todo.status}}" readonly="readonly">
+            <input type = 'hidden' name="card_id" value="{{todo.id}}" readonly="readonly">
             <input type="submit" name = 'action' value="Mark as Done"> 
             <input type="submit" name = 'action' value="Delete"> 
           </form>      
@@ -23,11 +23,11 @@
       <br>
       <div class="col-md-12">
         <h2>Current Completed Items from Trello:</h2>
-          {% for todo in cards_from_done %}
+          {% for todo in todos.done_items %}
             <form action="{{url_for('Items_To_Do')}}" method="POST">
               <input type ='text' name='card_name' value="{{todo.title}}" readonly="readonly">
-              <input type = 'text' name="list_id" value="{{todo.list_id}}" readonly="readonly">
-              <input type = 'text' name ="card_id" value="{{todo.id}}" readonly="readonly">
+              <input type = 'text' name="status" value="{{todo.status}}" readonly="readonly">
+              <input type = 'hidden' name="card_id" value="{{todo.id}}" readonly="readonly">
               <input type="submit" name = 'action' value="Mark as To Do"> 
               <input type="submit" name = 'action' value="Delete"> 
             </form>
