+++ view_model.py
@@ -8,17 +8,12 @@ class ViewModel:
 
     @property
     def to_do_items(self):
-        return [item for item in _items if item.status == "To Do"]
+        return [item for item in self._items if item.status == "To Do"]
         
-
     @property
     def doing_items(self):
-        for item in self._items:
-            if item.status == 'Doing':
-        return item
+        return [item for item in self._items if item.status == "Pending"]
 
     @property
     def done_items(self):
-        for item in self._items:
-            if item.status == 'Done':
-        return item
+        return [item for item in self._items if item.status == "Done"]
